Fixed some bugs related to reading 32 bit memory addresses that should have been read as 64 bit memory addresses.

+++ src/Poe/Elements/EntityLabel.cs
@@ -11,7 +11,7 @@ namespace PoeHUD.Poe.Elements
                 {
                     return "";
                 }
-                return LabelLen >= 8 ? M.ReadStringU(M.ReadInt(Address + 0xBB8), LabelLen * 2) : M.ReadStringU(Address + 0xBC8, LabelLen * 2);
+                return LabelLen >= 8 ? M.ReadStringU(M.ReadLong(Address + 0xBB8), LabelLen * 2) : M.ReadStringU(Address + 0xBC8, LabelLen * 2);
             }
         }
     }
